esys: Add support for more symmetric algorithms
This commit introduces a pair of generic symmetric
encryption/decryption callbacks to replace the algorithm-specific
callbacks, similar to the callbacks for hashes and hmacs.

When both the new and old callbacks are supplied, only the new ones
will be used. The existing requirement to supply callbacks for AES is
weakened to only requiring one pair of symmetric encryption/decryption
callbacks be supplied.

Both the OpenSSL and mbedtls backends now supply the new
callbacks. OpenSSL can handle all symmetric algorithms in the TCG
registry, assuming it is built with support for them. The mbedtls
backend only handles AES and Camellia, as mbedtls doesn't support SM4
or 3DES-CFB.

Signed-off-by: Fergus Dall <sidereal@google.com>